Refactors support page to use server actions
