+++ app/(gcforms)/[locale]/(support)/contact/page.tsx
@@ -1,7 +1,9 @@
 import { serverTranslation } from "@i18n";
 import { Metadata } from "next";
-import { Contact } from "./components/client/Contact";
-import { Success } from "./components/server/Success";
+import { ContactForm } from "./components/client/ContactForm";
+import { Success } from "../components/server/Success";
+import { Alert } from "@clientComponents/globals";
+import Link from "next/link";
 
 export async function generateMetadata({
   params: { locale },
@@ -19,14 +21,31 @@ export default async function Page({
 }: {
   searchParams: { success?: string };
 }) {
-  const { t } = await serverTranslation(["form-builder"]);
+  const {
+    t,
+    i18n: { language },
+  } = await serverTranslation(["form-builder"]);
 
   return (
     <>
       {success === undefined ? (
         <>
           <h1>{t("contactus.title")}</h1>
-          <Contact />
+          <p className="mb-6 mt-[-2rem] text-[1.6rem]">{t("contactus.useThisForm")}</p>
+          <p className="mb-14">
+            {t("contactus.gcFormsTeamPart1")}{" "}
+            <Link href={`https://www.canada.ca/${language}/contact.html`}>
+              {t("contactus.gcFormsTeamLink")}
+            </Link>{" "}
+            {t("contactus.gcFormsTeamPart2")}
+          </p>
+          <Alert.Warning title={t("contactus.needSupport")} role="note">
+            <p>
+              {t("contactus.ifYouExperience")}{" "}
+              <Link href={`/support`}>{t("contactus.supportFormLink")}</Link>.
+            </p>
+          </Alert.Warning>
+          <ContactForm />
         </>
       ) : (
         <>
