+++ app/(gcforms)/[locale]/(support)/support/page.tsx
@@ -1,7 +1,9 @@
-import { Support } from "./Support";
 import { serverTranslation } from "@i18n";
-
 import { Metadata } from "next";
+import { SupportForm } from "./components/client/SupportForm";
+import { Success } from "../components/server/Success";
+import Link from "next/link";
+import { Alert } from "@clientComponents/globals";
 
 export async function generateMetadata({
   params: { locale },
@@ -14,6 +16,41 @@ export async function generateMetadata({
   };
 }
 
-export default async function Page() {
-  return <Support />;
+export default async function Page({
+  searchParams: { success },
+}: {
+  searchParams: { success?: string };
+}) {
+  const {
+    t,
+    i18n: { language },
+  } = await serverTranslation(["form-builder"]);
+
+  return (
+    <>
+      {success === undefined ? (
+        <>
+          <h1>{t("support.title")}</h1>
+          <p className="mb-6 mt-[-2rem] text-[1.6rem]">{t("support.useThisForm")}</p>
+          <p className="mb-14">
+            {t("support.gcFormsTeamPart1")}{" "}
+            <Link href={`https://www.canada.ca/${language}/contact.html`}>
+              {t("support.gcFormsTeamLink")}
+            </Link>{" "}
+            {t("support.gcFormsTeamPart2")}
+          </p>
+          <Alert.Warning title={t("support.lookingForADemo")} role="note">
+            <p>
+              {t("support.ifYouWouldLike")} <Link href={`/contact`}>{t("support.contactUs")}</Link>.
+            </p>
+          </Alert.Warning>
+          <SupportForm />
+        </>
+      ) : (
+        <>
+          <Success />
+        </>
+      )}
+    </>
+  );
 }
